Isolate notification types
