+++ src/models/index.ts
@@ -8,7 +8,6 @@
 
 export * from './actionable';
 export * from './auto-extractor';
-export * from './broadcasted-notification';
 export * from './dashboard';
 export * from './entry';
 export * from './file';
@@ -30,7 +29,6 @@ export * from './search-module';
 export * from './search2';
 export * from './system-settings';
 export * from './tag';
-export * from './targeted-notification';
 export * from './template';
 export * from './timeframe';
 export * from './user';
