+++ src/models/notification/raw-broadcasted-notification.ts
@@ -16,17 +16,3 @@ export interface RawBroadcastedNotification {
 	Msg: string;
 	Origin: string;
 }
-
-export interface BroadcastedNotification {
-	id: string;
-	customID?: string;
-	type: 'broadcasted';
-	message: string;
-
-	sentDate: string; // Timestamp eg. '2019-04-22T21:44:01.776942432Z'
-	expirationDate: string; // Timestamp eg. '2019-04-23T03:44:01.776918756-06:00'
-	ignoreUntilDate: string; // Timestamp eg. '0001-01-01T00:00:00Z'
-
-	origin: string;
-	senderID: string;
-}
